Move Domain.kt to browser-api
