+++ app/src/androidTest/java/com/duckduckgo/app/trackerdetection/db/TdsTrackerDaoTest.kt
@@ -18,10 +18,10 @@ package com.duckduckgo.app.trackerdetection.db
 
 import androidx.room.Room
 import androidx.test.platform.app.InstrumentationRegistry
+import com.duckduckgo.app.browser.Domain
 import com.duckduckgo.app.global.db.AppDatabase
 import com.duckduckgo.app.trackerdetection.model.Action.BLOCK
 import com.duckduckgo.app.trackerdetection.model.TdsTracker
-import com.duckduckgo.common.Domain
 import org.junit.After
 import org.junit.Assert.assertEquals
 import org.junit.Assert.assertTrue
