+++ app/src/main/java/com/duckduckgo/app/browser/cookies/db/AuthCookiesAllowedDomainsRepository.kt
@@ -16,8 +16,8 @@
 
 package com.duckduckgo.app.browser.cookies.db
 
+import com.duckduckgo.app.browser.UriString
 import com.duckduckgo.common.utils.DispatcherProvider
-import com.duckduckgo.common.utils.UriString
 import javax.inject.Inject
 import kotlinx.coroutines.withContext
