+++ app/src/main/java/com/duckduckgo/app/fire/fireproofwebsite/data/FireproofWebsiteRepositoryImpl.kt
@@ -18,10 +18,10 @@ package com.duckduckgo.app.fire.fireproofwebsite.data
 
 import android.net.Uri
 import androidx.lifecycle.LiveData
+import com.duckduckgo.app.browser.UriString
 import com.duckduckgo.app.browser.favicon.FaviconManager
 import com.duckduckgo.app.fire.FireproofRepository
 import com.duckduckgo.common.utils.DispatcherProvider
-import com.duckduckgo.common.utils.UriString
 import com.duckduckgo.di.scopes.AppScope
 import com.squareup.anvil.annotations.ContributesBinding
 import dagger.Lazy
