+++ app/src/main/java/com/duckduckgo/app/trackerdetection/TrackerDetector.kt
@@ -19,6 +19,7 @@ package com.duckduckgo.app.trackerdetection
 import android.net.Uri
 import androidx.annotation.VisibleForTesting
 import com.duckduckgo.adclick.api.AdClickManager
+import com.duckduckgo.app.browser.UriString.Companion.sameOrSubdomainPair
 import com.duckduckgo.app.privacy.db.UserAllowListDao
 import com.duckduckgo.app.trackerdetection.Client.ClientType.BLOCKING
 import com.duckduckgo.app.trackerdetection.db.WebTrackerBlocked
@@ -27,7 +28,6 @@ import com.duckduckgo.app.trackerdetection.model.Entity
 import com.duckduckgo.app.trackerdetection.model.TrackerStatus
 import com.duckduckgo.app.trackerdetection.model.TrackerType
 import com.duckduckgo.app.trackerdetection.model.TrackingEvent
-import com.duckduckgo.common.utils.UriString.Companion.sameOrSubdomainPair
 import com.duckduckgo.di.scopes.AppScope
 import com.duckduckgo.privacy.config.api.ContentBlocking
 import com.duckduckgo.privacy.config.api.TrackerAllowlist
