+++ autofill/autofill-impl/src/main/java/com/duckduckgo/autofill/impl/RealAutofill.kt
@@ -16,9 +16,9 @@
 
 package com.duckduckgo.autofill.impl
 
+import com.duckduckgo.app.browser.UriString.Companion.sameOrSubdomain
 import com.duckduckgo.autofill.api.Autofill
 import com.duckduckgo.autofill.store.feature.AutofillFeatureRepository
-import com.duckduckgo.common.utils.UriString.Companion.sameOrSubdomain
 import com.duckduckgo.di.scopes.AppScope
 import com.duckduckgo.privacy.config.api.UnprotectedTemporary
 import com.squareup.anvil.annotations.ContributesBinding
