+++ autofill/autofill-impl/src/main/java/com/duckduckgo/autofill/impl/email/remoteconfig/EmailProtectionInContextExceptionsImpl.kt
@@ -16,8 +16,8 @@
 
 package com.duckduckgo.autofill.impl.email.remoteconfig
 
+import com.duckduckgo.app.browser.UriString.Companion.sameOrSubdomain
 import com.duckduckgo.autofill.store.feature.email.incontext.EmailProtectionInContextFeatureRepository
-import com.duckduckgo.common.utils.UriString.Companion.sameOrSubdomain
 import com.duckduckgo.di.scopes.AppScope
 import com.squareup.anvil.annotations.ContributesBinding
 import dagger.SingleInstanceIn
