--- common/common-kotlin/build.gradle
@@ -1,38 +1,0 @@
-/*
- * Copyright (c) 2022 DuckDuckGo
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-plugins {
-    id 'java-library'
-    id 'kotlin'
-    id 'com.google.devtools.ksp' version "$ksp_version"
-}
-
-apply from: "$rootProject.projectDir/code-formatting.gradle"
-
-java {
-    sourceCompatibility = JavaVersion.VERSION_17
-    targetCompatibility = JavaVersion.VERSION_17
-}
-
-kotlin {
-    jvmToolchain(17)
-}
-
-dependencies {
-    implementation KotlinX.coroutines.core
-    implementation "com.squareup.moshi:moshi-kotlin:_"
-    ksp "com.squareup.moshi:moshi-kotlin-codegen:1.15.0"
-}
