+++ common/common-utils/build.gradle
@@ -42,8 +42,6 @@ dependencies {
     anvil project(path: ':anvil-compiler')
     implementation project(path: ':anvil-annotations')
     implementation project(path: ':di')
-    implementation project(path: ':common-kotlin')
-
 
     implementation KotlinX.coroutines.core
     implementation KotlinX.coroutines.android
