+++ privacy-config/privacy-config-impl/src/main/java/com/duckduckgo/privacy/config/impl/features/useragent/RealUserAgent.kt
@@ -16,7 +16,7 @@
 
 package com.duckduckgo.privacy.config.impl.features.useragent
 
-import com.duckduckgo.common.utils.UriString
+import com.duckduckgo.app.browser.UriString
 import com.duckduckgo.di.scopes.AppScope
 import com.duckduckgo.privacy.config.api.DefaultPolicy
 import com.duckduckgo.privacy.config.api.DefaultPolicy.CLOSEST
