+++ privacy-dashboard/privacy-dashboard-impl/src/main/java/com/duckduckgo/privacy/dashboard/impl/ui/RequestDataViewStateMapper.kt
@@ -17,6 +17,7 @@
 package com.duckduckgo.privacy.dashboard.impl.ui
 
 import androidx.core.net.toUri
+import com.duckduckgo.app.browser.UriString
 import com.duckduckgo.app.global.model.Site
 import com.duckduckgo.app.trackerdetection.model.Entity
 import com.duckduckgo.app.trackerdetection.model.TrackerStatus
@@ -27,7 +28,6 @@ import com.duckduckgo.app.trackerdetection.model.TrackerStatus.SAME_ENTITY_ALLOW
 import com.duckduckgo.app.trackerdetection.model.TrackerStatus.SITE_BREAKAGE_ALLOWED
 import com.duckduckgo.app.trackerdetection.model.TrackerStatus.USER_ALLOWED
 import com.duckduckgo.app.trackerdetection.model.TrackingEvent
-import com.duckduckgo.common.utils.UriString
 import com.duckduckgo.common.utils.extractDomain
 import com.duckduckgo.common.utils.withScheme
 import com.duckduckgo.di.scopes.AppScope
