+++ request-filterer/request-filterer-impl/src/main/java/com/duckduckgo/request/filterer/impl/RequestFiltererImpl.kt
@@ -18,8 +18,8 @@ package com.duckduckgo.request.filterer.impl
 
 import android.webkit.WebResourceRequest
 import androidx.core.net.toUri
+import com.duckduckgo.app.browser.UriString
 import com.duckduckgo.common.utils.DispatcherProvider
-import com.duckduckgo.common.utils.UriString
 import com.duckduckgo.di.scopes.AppScope
 import com.duckduckgo.feature.toggles.api.FeatureToggle
 import com.duckduckgo.privacy.config.api.UnprotectedTemporary
